Updated Cloning Conversation
Cloning now deducts fee according to level

+++ scripts/conversation/medical_cloning_droid.py
@@ -20,16 +20,23 @@ def startConversation(core, actor, npc):
 	
 
 	if actor.getBuffByName('cloning_sickness'):
-		prose = ProsePackage('conversation/clone_droid', 's_6')
-		outOfBand = OutOfBand()
-		outOfBand.addProsePackage(prose)
-		convSvc.sendConversationMessage(actor, npc, outOfBand)
-		prose2 = ProsePackage('conversation/clone_droid', 's_8')
+	
+		level = int(actor.getLevel())
+		
+		if level <= 10:
+			cloningFee = 100
+		elif level >= 11:
+			cloningFee = (((actor.getLevel() + 15) * actor.getLevel())/2)
+		elif level == 90:
+			cloningFee = 5000
+			
+		convSvc.sendConversationMessage(actor, npc, OutOfBand.ProsePackage("@conversation/clone_droid:s_6", int(cloningFee)))
+		prose1 = ProsePackage('conversation/clone_droid', 's_8')
 		outOfBand2 = OutOfBand()
-		outOfBand2.addProsePackage(prose2)
-		prose3 = ProsePackage('conversation/clone_droid', 's_12')
+		outOfBand2.addProsePackage(prose1)
+		prose2 = ProsePackage('conversation/clone_droid', 's_12')
 		outOfBand3 = OutOfBand()
-		outOfBand3.addProsePackage(prose3)
+		outOfBand3.addProsePackage(prose2)
 		option1 = ConversationOption(outOfBand2, 0)
 		option2 = ConversationOption(outOfBand3, 1)
 
@@ -41,7 +48,7 @@ def startConversation(core, actor, npc):
 		convSvc.sendConversationOptions(actor, npc, options, handleFirstScreen)
 		return
 	else:
-		core.conversationService.sendStopConversation(actor, npc, 'conversation/clone_droid', 's_4')
+		core.conversationService.sendStopConversation(actor, npc, 'conversation/clone_droid', 's_14')
 		return	
 	return
 	
@@ -68,24 +75,48 @@ def handleFirstScreen(core, actor, npc, selection):
 
 		
 	if selection == 1:
-		core.buffService.removeBuffFromCreature(actor, actor.getBuffByName("cloning_sickness"));
-		core.conversationService.sendStopConversation(actor, npc, 'conversation/clone_droid', 's_14')
-		return
-	return		
+		if actor.getLevel() <= 10:
+			cloningFee = 100
+		elif actor.getLevel() >= 11:
+			cloningFee = (((actor.getLevel() + 15) * actor.getLevel())/2)
+		elif actor.getLevel() == 90:
+			cloningFee = 5000
+
+			
+		if actor.getCashCredits() >= cloningFee:
+			core.buffService.removeBuffFromCreature(actor, actor.getBuffByName("cloning_sickness"));
+			core.conversationService.sendStopConversation(actor, npc, 'conversation/clone_droid', 's_14')
+			actor.setCashCredits(actor.getCashCredits() - int(cloningFee))
+			return
+		elif actor.getCashCredits() <= cloningFee:
+			core.conversationService.sendStopConversation(actor, npc, 'conversation/clone_droid', 's_16')
+			return
+		return		
 	
 	
 def handleSecondScreen(core, actor, npc, selection):
 	convSvc = core.conversationService
 	
-	
+	if actor.getLevel() <= 10:
+		cloningFee = 100
+	elif actor.getLevel() >= 11:
+		cloningFee = (((actor.getLevel() + 15) * actor.getLevel())/2)
+	elif actor.getLevel() == 90:
+		cloningFee = 5000
 	
 	if selection == 0:
-		core.buffService.removeBuffFromCreature(actor, actor.getBuffByName("cloning_sickness"));
-		core.conversationService.sendStopConversation(actor, npc, 'conversation/clone_droid', 's_14')
+		if actor.getCashCredits() >= cloningFee:
+			core.buffService.removeBuffFromCreature(actor, actor.getBuffByName("cloning_sickness"));
+			core.conversationService.sendStopConversation(actor, npc, 'conversation/clone_droid', 's_14')
+			actor.setCashCredits(actor.getCashCredits() - int(cloningFee))
+			return
+		elif actor.getCashCredits() <= cloningFee:
+			core.conversationService.sendStopConversation(actor, npc, 'conversation/clone_droid', 's_16')
+			return
 		return
-	return
 	
 def endConversation(core, actor, npc):
 	core.conversationService.sendStopConversation(actor, npc, 'conversation/clone_droid', 's_4')
 	return
-	
+			
+
